Silence some REALLY STUPID gcc/clang warnings! (SIGH!)

+++ archlvl.c
@@ -405,7 +405,7 @@ FACILITY( DETECT_PGMINTLOOP,S370|ESA390|ZARCH, NONE, S370|ESA390|ZARCH, ALS0|ALS
 void init_als( REGS* regs )
 {
     int  i;
-    for (i=0; i < STFL_HERCBYSIZE; i++)
+    for (i=0; i < (int) STFL_HERCBYSIZE; i++)
         regs->facility_list[i] =
             sysblk.facility_list[ regs->arch_mode ][i];
 }
@@ -418,7 +418,7 @@ static void set_alslevel( int alslevel )
     FACTAB*  ft;
     int      i, j;
 
-    for(i = 0; i < STFL_HERCBYSIZE; i++)
+    for(i = 0; i < (int) STFL_HERCBYSIZE; i++)
         for(j = 0; j < GEN_MAXARCH; j++)
             sysblk.facility_list[j][i] = 0;
 
@@ -737,7 +737,7 @@ static BYTE update_facility( int argc, char* argv[] )
         && isdigit( *( argv[2] + 3))
         && sscanf(     argv[2] + 3, "%d%c", &bitno, &c ) == 1
         && bitno >= 0
-        && bitno <= STFL_HERCMAX
+        && bitno <= (int) STFL_HERCMAX
     )
     {
         force_facbit( bitno, enable, als );
@@ -838,7 +838,7 @@ int archlvl_cmd( int argc, char* argv[], char* cmdline )
                 && isdigit( *( argv[2] + 3 ))
                 && sscanf(     argv[2] + 3, "%d%c", &bitno, &c ) == 1
                 && bitno >= 0
-                && bitno <= STFL_HERCMAX
+                && bitno <= (int) STFL_HERCMAX
             )
             {
                 const char* name = NULL;
